Update Quiz to be VOICE Quiz

+++ master_to_parent2.py
@@ -78,7 +78,10 @@ def replace_idea_with_voice(parent_course_id):
     # Look for items in modules containing "IDEA" and remove them
     for module in modules:
         for module_item in module.get_module_items():
-            if "IDEA" in module_item.title:
+            if (
+                "IDEA" in module_item.title
+                and "extra credit" not in module_item.title.lower()
+            ):
                 module_item.delete()
 
     announcements = parent_course.get_discussion_topics(only_announcements=True)
@@ -103,7 +106,10 @@ def replace_idea_with_voice(parent_course_id):
 
     assignments = parent_course.get_assignments()
     for assignment in assignments:
-        if "IDEA" in assignment.name:
+        if (
+            "IDEA" in assignment.name
+            and "extra credit" not in module_item.title.lower()
+        ):
             assignment.delete()
 
     # In the module named “Instructor Resources” search for a page where the name contains “Course Support Materials” and rename that page “Instructors - READ ME”
@@ -120,21 +126,37 @@ def replace_idea_with_voice(parent_course_id):
         parent_course.syllabus_body.replace("IDEA", "VOICE")
         # parent_course.edit(course={"syllabus_body": syllabus})
 
-    assignments = parent_course.get_assignments()
-    for assignment in assignments:
-        if "VOICE" in assignment.name:
-            break
-    else:
-        if voice_assignment_group:
-            parent_course.create_assignment(
-                assignment={
-                    "name": "VOICE Course Evaluation",
-                    "published": True,
-                    "assignment_group_id": voice_assignment_group.id,
+    # 1. Find the Extra Credit Quiz
+    quizzes = parent_course.get_quizzes()
+
+    for quiz in quizzes:
+        if "idea survey quiz" in quiz.title.lower():
+            quiz.edit(
+                quiz={
+                    "title": "Extra Credit: VOICE Survey Quiz",
+                    "description": """
+<div class="cc-container">
+    <p>If you complete a survey for this course, you will be eligible for one point of extra credit on your overall grade. To receive the extra credit, you must perform both of the following steps:</p>
+    <p>STEP 1: <a href="/users/self/external_tools/410321">Complete the VOICE survey here</a></p>
+    <p>STEP 2: Confirm that you have completed the survey by taking this quiz.</p>
+    <p>More information about the VOICE Survey or if you experience any technical difficulties, please visit the <a href="https://elearning.champlain.edu/voice-survey-instructions/">CCO VOICE Survey Instructions</a>.</p>
+</div>""",
                 }
             )
-        else:
-            print("MISSING VOICE ASSIGNMENT GROUP FOR ", parent_course.name)
+            questions = quiz.get_questions()
+            question_count = 0
+            for question in questions:
+                question.edit(
+                    question={
+                        "question_text":"VOICE Question",
+                        "question_name": "<p>Did you complete the VOICE survey?</p>"
+
+                    }
+                    question_count += 1
+                    if question_count > 1:
+                        print(f"There are too many questions in the quiz! {parent_course.name} Quiz id: {quiz.id}")
+                )
+
 
 
 def migrate_every_master_to_parent(master_courses):
@@ -203,8 +225,8 @@ all_master_courses = canvas.get_account(master_account_id).get_courses()
 
 # UNCOMMENT WHEN YOU WANT TO RUN THE WHOLE SCRIPT
 # RUN THE FIRST LINE BY ITSELF THEN A FEW HOURS LATER RUN THE REST
-# new_parent_ids = migrate_master_to_parent_from_csv("parent_course_list.csv")
-# generate_spreadsheet(new_parent_ids)
+new_parent_ids = migrate_master_to_parent_from_csv("parent_course_list.csv")
+generate_spreadsheet(new_parent_ids)
 # STOP HERE AND WAIT FOR MIGRATIONS TO COMPLETE
 new_parent_ids = read_parent_ids_from_csv()
 remove_idea_from_all(new_parent_ids)
